backend/backend/requirements.txt
asgiref==3.8.1
boto3==1.35.31
botocore==1.35.31
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
Markdown==3.7
PyJWT==2.9.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
